+++ src/services/agents.service.js
@@ -1,7 +1,7 @@
 import z from 'zod';
 import * as agent_builder from '../builders/agent.builder.js';
 import Guard from '../utils/guard.util.js';
-
+import { generateToken } from '../utils/token.util.js';
 /**
  * Service that launchs the execution of agent creation.
  * @param {String} label label of the new agent to be created.
@@ -20,5 +20,7 @@ export const create = async function (
     type: z.string().min(2),
   });
   const data = Guard.validateProps(schema, props);
-  return await fns.agent_create(data);
-};
+  const agent = await fns.agent_create(data);
+  agent.token = await generateToken({ id_user: agent.id_agent, is_agent: true });
+  return agent;
+}